backend-ms1/src/test/java/ma/zyn/app/unit/service/impl/admin/app/AppStateAdminServiceImplTest.java
package ma.zyn.app.unit.service.impl.admin.app;

import ma.zyn.app.bean.core.app.AppState;
import ma.zyn.app.dao.facade.core.app.AppStateDao;
import ma.zyn.app.service.impl.admin.app.AppStateAdminServiceImpl;

import java.util.*;
import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.stream.Collectors;
import java.util.stream.IntStream;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import java.time.LocalDateTime;



import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;


@SpringBootTest
class AppStateAdminServiceImplTest {

    @Mock
    private AppStateDao repository;
    private AutoCloseable autoCloseable;
    private AppStateAdminServiceImpl underTest;



    @BeforeEach
    void setUp() {
        autoCloseable = MockitoAnnotations.openMocks(this);
        underTest = new AppStateAdminServiceImpl(repository);
    }

    @AfterEach
    void tearDown() throws Exception {
        autoCloseable.close();
    }

    @Test
    void canGetAllAppState() {
         //when
        underTest.findAll();
        verify(repository).findAll();
    }

    @Test
    void itShouldSaveAppState() {
        // Given
        AppState toSave = constructSample(1);
        when(repository.save(toSave)).thenReturn(toSave);

        // When
        underTest.create(toSave);

        // Then
        verify(repository).save(toSave);
    }

    @Test
    void itShouldDeleteAppState() {
        // Given
        Long idToDelete = 1L;
        when(repository.existsById(idToDelete)).thenReturn(true);

        // When
        underTest.deleteById(idToDelete);

        // Then
        verify(repository).deleteById(idToDelete);
    }
    @Test
    void itShouldGetAppStateById() {
        // Given
        Long idToRetrieve = 1L; // Example AppState ID to retrieve
        AppState expected = new AppState(); // You need to replace AppState with your actual class
        expected.setId(idToRetrieve);
        when(repository.findById(idToRetrieve)).thenReturn(java.util.Optional.of(expected));

        // When
        AppState result = underTest.findById(idToRetrieve);

        // Then
        assertEquals(expected, result);
    }
	
	private AppState constructSample(int i) {
		AppState given = new AppState();
        given.setCode("code-"+i);
        given.setLibelle("libelle-"+i);
        given.setStyle("style-"+i);
        return given;
    }

}
